made single card component and also a card list component that maps over the cards
