+++ src/App.css
@@ -9,15 +9,3 @@
 }
 
 
- .title {
-  display: flex;
-  flex-direction: column;;
-} 
-
-.MuiCardHeader-content css-1qbkelo-MuiCardHeader-content {
-
-}
-
-.title1 {
-  padding-top: 40px;
-}
