+++ src/App.js
@@ -1,5 +1,6 @@
 import './App.css';
 import RecCardList from './Components/RecCardList';
+import SingleCard from './Components/SingleCard';
 
 function App() {
   return (
